Android: Fix UI is scaled smaller than before
The 413593183bbb1137fdc784d98c171d67a167bb32 patch changed the way
how the display metrics are retrieved. By doing so, it was found that
the previous way retrieved the scaledDensity always equal to density.
It is intentional for scaledDensity to be dependent on the font scale
chosen by the user. However, this change altered not only the font scale
but also the layout. This patch will make the layout dependent on the
density instead of the scaledDensity and normalize the way the display
metrics are retrieved among Android versions.
Currently, the fontScale is ignored, QTBUG-109566 will track future
developments.

Fixes: QTBUG-109026
Change-Id: I6adacd17583cbe9bee368af35c50b780872ab222
Reviewed-by: Ville Voutilainen <ville.voutilainen@qt.io>
(cherry picked from commit 99893a914a821567e10935ffb8be24df7147ccd9)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>