Move QApplication::autoSipEnabled() to public scope
The method is not meant be a slot.

warning: getter QApplication::autoSipEnabled possibly mismarked as a
slot [clazy-const-signal-or-slot]

[ChangeLog][QtWidgets][Potentially Source-Incompatible Changes]
bool QApplication::autoSipEnabled() is no longer a slot. Code such as
using QObject::connect() to connect to it as a slot or accessing it
through meta object system will have to be rewritten.

Change-Id: I892a51120478f3c90e5834c8e3e00bc836b84c52
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit 03ed57afa0c5af5076583dd4e415e54983e7e0a8)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>